- Disable compilation for the DuckType benchmarks (so we can skip them by default). - Refactor the ProxyActivator class and make it public.
