+++ benchmarks/Benchmarks.Trace/Benchmarks.Trace.csproj
@@ -18,6 +18,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="DuckTyping\DuckTypeMethodBenchmark.cs" />
+    <Compile Remove="DuckTyping\DuckTypeValueTypeFieldBenchmark.cs" />
+    <Compile Remove="DuckTyping\DuckTypeValueTypePropertyBenchmark.cs" />
+    <Compile Remove="DuckTyping\ObscureObject.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
   </ItemGroup>
@@ -42,6 +49,13 @@
 
   <ItemGroup>
     <None Remove="microbenchmarks\.gitignore" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="DuckTyping\DuckTypeMethodBenchmark.cs" />
+    <None Include="DuckTyping\DuckTypeValueTypeFieldBenchmark.cs" />
+    <None Include="DuckTyping\DuckTypeValueTypePropertyBenchmark.cs" />
+    <None Include="DuckTyping\ObscureObject.cs" />
   </ItemGroup>
 
 </Project>
